requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-appconf==1.0.5
django-braces==1.14.0
django-ckeditor==6.1.0
django-googledrive-storage==1.5.2
django-imagefit==0.6.2
django-js-asset==1.2.2
django-shortcuts==1.6
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
waitress==2.0.0
